Changed
Interesting info from
https://github.com/rhinstaller/shim/blob/master/README.fallback

When you boot removable media, it'll be in \EFI\BOOT , but fallback.efi
won't be there, so it goes ahead and boots the normal bootloader
(grubx64.efi).  When you boot a normal system through a boot variable,
the boot variable is configured to start \EFI\fedora\shim.efi (or
whatever your distro's EFI directory is.)  In that case it won't try to
invoke fallback.  But if the boot variables are missing or corrupted,
the firmware will eventually try to boot the hard disk as removable
media.  In that case, it'll invoke \EFI\BOOT\BOOTX64.EFI (or whatever
filename is right for your architecture.)  In that case it'll be in
\EFI\BOOT, so it'll check for fallback.efi , and it'll find it and run
it.  When it runs, fallback will look for every directory in \EFI\ with
a BOOT${ARCH}.CSV in it, or BOOT.CSV if that's not found.  It'll parse that,
and create new boot variables from what it finds.  Then it'll try to boot one
of them.